Refractor project structure
