+++ src/redux/store.js
@@ -1,0 +1,1 @@
+// Redux State management tool
