+++ src/utils/constants.js
@@ -1,0 +1,1 @@
+export const API_ROOT = 'http://localhost:8017'
